Move redirect url generation logic
